requirements.txt
alembic==1.16.5
bcrypt==5.0.0
blinker==1.9.0
cachelib==0.13.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.6
msgspec==0.19.0
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
reportlab==4.4.4
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
Werkzeug==3.1.3
